split relation_scoping_test.rb's default scoping tests into another file
